Create the DOM tree for the page layout
Version 1, with refinements probably needed.

Also need to find something to use as the "images". A scaled down version of the
grid used elsewhere just might be the answer. Created three different sizes for
showing the results of images which don't match the aspect ratio, exactly, of
the container for them.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>